fix: sidebar code icon reset bug (#9307) (#9333)
* fix: 修复侧边栏重置时 Code 图标消失的问题 (#9307)

问题原因：
- types/index.ts 中的 SidebarIcon 类型定义缺少 'code_tools'
- 存在重复的类型定义和常量定义导致不一致

修复内容：
- 在 types/index.ts 的 SidebarIcon 类型中添加 'code_tools'
- 删除 minapps.ts 中重复的 DEFAULT_SIDEBAR_ICONS 常量
- 统一从 @renderer/types 导入 SidebarIcon 类型
- 删除 settings.ts 中重复的 SidebarIcon 类型定义

这确保了在导航栏设置为左侧时，点击侧边栏设置的重置按钮后，
Code 图标能够正确显示。

* refactor: 将侧边栏配置移至 config 目录

根据 code review 建议，将侧边栏相关配置从 store/settings.ts
移动到 config/sidebar.ts，使配置管理更加清晰。

改动内容：
- 创建 config/sidebar.ts 存放侧边栏配置常量
- 更新相关文件的导入路径
- 在 settings.ts 中重新导出以保持向后兼容
- 添加 REQUIRED_SIDEBAR_ICONS 常量便于未来扩展

这个改动保持了最小化原则，不影响现有功能。
