+++ src/renderer/src/store/minapps.ts
@@ -1,16 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { DEFAULT_MIN_APPS } from '@renderer/config/minapps'
-import { MinAppType, SidebarIcon } from '@renderer/types'
-
-export const DEFAULT_SIDEBAR_ICONS: SidebarIcon[] = [
-  'assistants',
-  'agents',
-  'paintings',
-  'translate',
-  'minapp',
-  'knowledge',
-  'files'
-]
+import { MinAppType } from '@renderer/types'
 
 export interface MinAppsState {
   enabled: MinAppType[]
