+++ src/renderer/src/store/settings.ts
@@ -1,5 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { TRANSLATE_PROMPT } from '@renderer/config/prompts'
+import { DEFAULT_SIDEBAR_ICONS } from '@renderer/config/sidebar'
 import {
   ApiServerConfig,
   AssistantsSortType,
@@ -10,6 +11,7 @@ import {
   OpenAISummaryText,
   PaintingProvider,
   S3Config,
+  SidebarIcon,
   ThemeMode,
   TranslateLanguageCode
 } from '@renderer/types'
@@ -21,26 +23,8 @@ import { RemoteSyncState } from './backup'
 
 export type SendMessageShortcut = 'Enter' | 'Shift+Enter' | 'Ctrl+Enter' | 'Command+Enter' | 'Alt+Enter'
 
-export type SidebarIcon =
-  | 'assistants'
-  | 'agents'
-  | 'paintings'
-  | 'translate'
-  | 'minapp'
-  | 'knowledge'
-  | 'files'
-  | 'code_tools'
-
-export const DEFAULT_SIDEBAR_ICONS: SidebarIcon[] = [
-  'assistants',
-  'agents',
-  'paintings',
-  'translate',
-  'minapp',
-  'knowledge',
-  'files',
-  'code_tools'
-]
+// Re-export for backward compatibility
+export { DEFAULT_SIDEBAR_ICONS }
 
 export interface NutstoreSyncRuntime extends RemoteSyncState {}
